Restructured tracking files
